+++ game/src/main/java/org/ttarena/arena_game/document/GameSession.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.annotation.Version;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -21,6 +22,9 @@ public class GameSession {
 
     @Id
     private String id;
+
+    @Version
+    private Long version;
 
     private List<Participant> participants;