+++ map/src/main/java/org/ttarena/arena_map/document/GameMap.java
@@ -4,6 +4,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.annotation.Version;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.ttarena.arena_map.model.HexCoordinate;
@@ -21,6 +22,9 @@ public class GameMap {
 
     @Id
     private String id;
+
+    @Version
+    private Long version;
 
     @Indexed
     private String ownerId;